test_environment/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
astunparse==1.6.3
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
colorthief==0.2.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
dataclasses==0.6
fastapi==0.115.6
filelock==3.16.1
fire==0.7.0
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
imantics==0.1.12
imgaug==0.4.0
iopath==0.1.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
kiwisolver==1.4.8
labelme2coco==0.2.6
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
pixellib==0.7.1
platformdirs==4.3.6
portalocker==3.0.0
protobuf==5.29.2
pybboxes==0.1.6
pydantic==2.10.4
pydantic_core==2.27.2
pydot==3.0.3
Pygments==2.18.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
sahi==0.11.20
scikit-image==0.25.0
scipy==1.14.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminaltables==3.1.10
tifffile==2024.12.12
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.0
xmljson==0.2.1
yacs==0.1.8
